requirements.txt
numpy
scipy
pandas
Pillow
langchain
streamlit
llama-index
streamlit-modal
beautifulsoup4